Update anonchat_msg.js
I have included the capability to turn on and off the AnonChat feature by using the command "xac on". This command can be used without any prefix or explicit call. However, it is important to note that the command relies on a file called "anonchat_auto.js", which must be saved in the "onMessage" folder. It is also crucial that the names of the commands remain as "anonchat_msg.js" and "anonchat_auto.js", as they are used to read the global object. Changing the names will cause the feature to malfunction.